requirements.txt
# Core Dependencies - Upgraded for Gemini 3 support
# Todos os pacotes LangChain precisam ser compatíveis entre si (ecossistema v1.x)
langchain-google-genai>=4.0.0
langchain-community>=0.4.0
langchain-openai>=0.3.0
langgraph>=0.3.0
openai>=1.54.4
langchain-anthropic>=0.4.0
anthropic>=0.28.0
google-genai>=1.0.0

# HTTP
httpx>=0.23.0

# Web Framework
fastapi==0.115.4
uvicorn[standard]==0.32.0
pydantic>=2.7.4,<3.0.0
pydantic-settings>=2.6.0

# HTTP & API
requests>=2.32.5
typesense>=0.21.0

# Database & Storage
redis==4.6.0  # ARQ requer redis<5
psycopg==3.2.12
psycopg2-binary==2.9.10  # Para compatibilidade com código existente

# AI & ML
cohere==4.47

# Utilities
python-dotenv==1.0.0
pytz==2024.1

# Queue & Background Tasks
arq==0.26.0
supervisor==4.2.5

# Optional: Logging & Monitoring
python-json-logger==2.0.7

# Re-ranking removido - O Analista (LLM) decide o melhor produto
apscheduler==3.10.4
